+++ F-BJV24001115-synergy7-mjp-binfood-ch5/src/main/java/com/marceljsh/binarfud/payload/request/ProductUpdateRequest.java
@@ -8,6 +8,8 @@ import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ProductUpdateRequest {
@@ -20,5 +22,5 @@ public class ProductUpdateRequest {
   private String name;
 
   @PositiveOrZero(message = "price must be zero or greater")
-  private long price;
+  private BigDecimal price;
 }
